Filter all events thru 'jq' to provide Results page.
This is a first step towards getting the interface going. All Uchiwa events are
downloaded into a EventsData array, which is then re-marshalled and filtered through
Query.Filter -- which is just a 'jq boolean expression' placed into a jq exec.

This obvs needs a lot of work, but it's quite a nice that by calling 'jq' we can effectively
learn good JQ query syntax.