Pull in template logic from go-jira; implement help page
At some point this template handling code will probably be split
out into a library. It's not practical to just import go-jira, and
frankly this would be confusing and unneccessary.

So, C&P the code in, tweak slightly and remove unnecessary code/imports

Then, make sure it works by enabling the help page. Which works, woohoo!