Get ShowResultDetail page working; ditch word-wrap code
This enables the ShowDetailPage, just with the debug (toJson) template
for now. It's working, as is Refresh (though the backend data isn't refreshed yet)

I don't think there's a need for the word-wrapping functionality in tessen, at least
not the very JIRA specific code that we have, so remove that. Less code FTW.