Fix bug where a lack of query results caused page to break
Previously, on an empty result list, the page would be blank and
navigation would halt. This bug is technically still there, but by
ensuring that we never have an empty result set (by adding a dummy line
giving the user information that there were no results), we prevent it
from surfacing.