Initial HTTP dashboard support
Run a web dashboard on given listen address:port, when called with --listen addr:port

Any query with 'dashboard: true' in the config.yml will be displayed as a traffic-light
in a single page.

This is super shonky and needs a refactor, it is a first iteration to get this operational.