Implement Sources for data
This opens the door for (eg) PagerDuty integration, and hopefully other
data source types.

Largely an exercise in bringing the cached JSON from the source's provider
API into the Source struct, and collecting from each one.

This also enables collection from multiple uchiwa sources, though i've not tested
it heavily.

There's more to do to make this generic, but it's a reasonable start.

Technically, there is a bunch of work here to handle the fact that we store
everything as an interface{} now, which i still think is butt ugly. I don't have a
better solution right now though.

Some key changes:

* It's not possible now to specify --endpoint as a cli opt, as we have and endpoint
  per source. Same goes for UCHIWA_ENDPOINT
* Similarly, baseQueries is somewhat redundant. Shame for newbs, there may be a way
  of doing something similar.
* Removed the 'list' cli command and --query option, they are not compatible with
  the use of multiple sources, as above