Add a pagerduty provider!
This doesn't yet handle pagerduty API token, it's just a POC to get the
data side working with a file:// endpoint against test data.

No proper templating yet, just using debug for the display.

But it works!

Also includes a bit more of a re-org of some uchiwa code into the uchiwa
provider