Bodge http_dashboard to handle pagerduty provider
This is butt ugly, but it gets the dashboard working again.

Treat Pagerduty 'triggered' events as critical, and 'acknowledged' as 'warning'. There
is no parallel with 'Unknown', so just count those as zero.

I really don't like this, but haven't thought of an elegant way of handling the template
and CSS classes et al yet. So this gets it operational, which is The Main Thing.

Meh.